+++ projects/demo/src/app/app.component.ts
@@ -1,14 +1,13 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 import { NgxButtonComponent } from 'ngx-zen/button';
+import { NgxAlertComponent } from 'ngx-zen/alert';
 
 @Component({
   selector: 'app-root',
   standalone: true,
-  imports: [RouterOutlet, NgxButtonComponent],
+  imports: [RouterOutlet, NgxButtonComponent, NgxAlertComponent],
   templateUrl: './app.component.html',
-  styleUrl: './app.component.scss'
+  styleUrl: './app.component.scss',
 })
-export class AppComponent {
-
-}
+export class AppComponent {}